Move convertActionParameterArrayToMap to ConfigExtension
