+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/register/RegisterViewModel.kt
@@ -48,7 +48,7 @@ import org.smartregister.fhircore.engine.util.SharedPreferenceKey
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
 import org.smartregister.fhircore.quest.data.register.RegisterPagingSource
 import org.smartregister.fhircore.quest.data.register.model.RegisterPagingSourceState
-import org.smartregister.fhircore.quest.util.convertActionParameterArrayToMap
+import org.smartregister.fhircore.quest.util.extensions.convertActionParameterArrayToMap
 
 @HiltViewModel
 class RegisterViewModel
