+++ android/quest/src/test/java/org/smartregister/fhircore/quest/util/IterableUtilTest.kt
@@ -19,8 +19,6 @@ package org.smartregister.fhircore.quest.util
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
-import org.smartregister.fhircore.engine.domain.model.ActionParameter
-import org.smartregister.fhircore.engine.domain.model.ActionParameterType
 
 class IterableUtilTest {
   private lateinit var map: Map<String, Int?>
@@ -52,36 +50,5 @@ class IterableUtilTest {
   fun testNonNullGetOrDefaultShouldReturnValueIfKeyExists() {
     val defaultValue = 1
     Assert.assertEquals(24, nonNullGetOrDefault(map, "Vanilla", defaultValue))
-  }
-
-  @Test
-  fun testConvertActionParameterArrayToMapShouldReturnEmptyMapIfEmpty() {
-    Assert.assertEquals(emptyMap<String, String>(), convertActionParameterArrayToMap(null))
-  }
-
-  @Test
-  fun testConvertActionParameterArrayToMapShouldReturnEmptyMapIfNoParamData() {
-    val array = arrayOf(ActionParameter(key = "k", value = "v"))
-    Assert.assertEquals(emptyMap<String, String>(), convertActionParameterArrayToMap(array))
-  }
-
-  @Test
-  fun testConvertActionParameterArrayToMapShouldReturnEmtpyMapIfArrayIsEmpty() {
-    val array = emptyArray<ActionParameter>()
-    Assert.assertEquals(emptyMap<String, String>(), convertActionParameterArrayToMap(array))
-  }
-
-  @Test
-  fun testConvertActionParameterArrayToMapShouldReturnEmtpyMapValue() {
-    val array =
-      arrayOf(ActionParameter(key = "k", value = "", paramType = ActionParameterType.PARAMDATA))
-    Assert.assertEquals("", convertActionParameterArrayToMap(array)["k"])
-  }
-
-  @Test
-  fun testConvertActionParameterArrayToMapShouldReturnMapIfParamData() {
-    val array =
-      arrayOf(ActionParameter(key = "k", value = "v", paramType = ActionParameterType.PARAMDATA))
-    Assert.assertEquals(mapOf("k" to "v"), convertActionParameterArrayToMap(array))
   }
 }
